requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
alembic==0.9.3
async-timeout==4.0.1
attrs==21.2.0
bleach==2.0.0
blinker==1.4
certifi==2021.10.8
charset-normalizer==2.0.7
click==6.7
dominate==2.3.1
Flask==0.12.2
Flask-Bootstrap==3.3.7.1
Flask-HTTPAuth==3.2.3
Flask-Login==0.4.0
Flask-Mail==0.9.1
Flask-Migrate==2.0.4
Flask-Moment==0.5.1
Flask-PageDown==0.2.2
Flask-Script==2.0.6
Flask-SQLAlchemy==2.2
Flask-WTF==0.14.2
frozenlist==1.2.0
html5lib==0.999999999
idna==3.3
itsdangerous==0.24
Jinja2==2.9.6
Mako==1.0.7
Markdown==2.6.8
MarkupSafe==1.1.1
meraki==1.15.0
multidict==5.2.0
python-dateutil==2.6.1
python-dotenv==0.6.5
python-editor==1.0.3
requests==2.26.0
six==1.10.0
SQLAlchemy==1.1.11
typing-extensions==4.0.0
urllib3==1.26.7
visitor==0.1.3
webencodings==0.5.1
Werkzeug==0.12.2
WTForms==2.1
yarl==1.7.2
